removed useless components and renamed login components
