--- src/components/InfoBar/InfoBar.scss
@@ -1,7 +1,0 @@
-.infoBar {
-  font-variant: italics;
-}
-
-.time {
-  margin: 0 30px;
-}
