--- src/components/SurveyForm/SurveyForm.scss
@@ -1,42 +1,0 @@
-.inputGroup {
-  position: relative;
-}
-
-.flags {
-  position: absolute;
-  right: 20px;
-  top: 7px;
-  & > * {
-    margin: 0 2px;
-    width: 20px;
-    height: 20px;
-    border-radius: 20px;
-    box-shadow: 1px 1px 3px rgba(0, 0, 0, 0.4);
-    color: white;
-    float: right;
-    text-align: center;
-  }
-  .active {
-    background: linear-gradient(#cc0, #aa0);
-    color: black;
-  }
-  .dirty {
-    background: linear-gradient(#090, #060);
-  }
-  .visited {
-    background: linear-gradient(#009, #006);
-  }
-  .touched {
-    background: linear-gradient(#099, #066);
-  }
-}
-
-.radioLabel {
-  margin: 0 25px 0 5px;
-}
-
-.cog {
-  position: absolute;
-  left: 0;
-  top: 10px;
-}
