+++ src/components/index.js
@@ -5,10 +5,7 @@
  *
  */
 
-export CounterButton from './CounterButton/CounterButton';
 // export FacebookLogin from './FacebookLogin/FacebookLogin';
-export GithubButton from './GithubButton/GithubButton';
-export InfoBar from './InfoBar/InfoBar';
 // export LoginForm from './LoginForm/LoginForm';
 // export MiniInfoBar from './MiniInfoBar/MiniInfoBar';
 export Notifs from './Notifs/Notifs';
