--- src/containers/Chat/Loadable.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import Loadable from 'react-loadable';
-
-const ChatLoadable = Loadable({
-  loader: () => import('./Chat' /* webpackChunkName: 'chat' */),
-  loading: () => <div>Loading</div>
-});
-
-export default ChatLoadable;
