--- src/containers/ChatFeathers/Loadable.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import Loadable from 'react-loadable';
-
-const ChatFeathersLoadable = Loadable({
-  loader: () => import('./ChatFeathers' /* webpackChunkName: 'chat-feathers' */),
-  loading: () => <div>Loading</div>
-});
-
-export default ChatFeathersLoadable;
