--- src/containers/Widgets/Loadable.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import Loadable from 'react-loadable';
-
-const WidgetsLoadable = Loadable({
-  loader: () => import('./Widgets' /* webpackChunkName: 'widgets' */),
-  loading: () => <div>Loading</div>
-});
-
-export default WidgetsLoadable;
