+++ src/containers/login/loadable.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Loadable from 'react-loadable';
 
 const LoginLoadable = Loadable({
-  loader: () => import('./Login' /* webpackChunkName: 'login' */),
+  loader: () => import('./login' /* webpackChunkName: 'login' */),
   loading: () => <div>Loading</div>
 });
