+++ src/routes.js
@@ -2,13 +2,9 @@ import { routerActions } from 'react-router-redux';
 import { connectedReduxRedirect } from 'redux-auth-wrapper/history4/redirect';
 import { App, NotFound } from 'containers';
 import About from 'containers/about/about';
-import Chat from 'containers/Chat/Loadable';
-import ChatFeathers from 'containers/ChatFeathers/Loadable';
-import Login from 'containers/Login/Loadable';
+import Login from 'containers/login/loadable';
 import LoginSuccess from 'containers/LoginSuccess/Loadable';
 import Register from 'containers/Register/Loadable';
-import Survey from 'containers/Survey/Loadable';
-import Widgets from 'containers/Widgets/Loadable';
 
 import Home from './containers/home/home';
 
@@ -51,7 +47,6 @@ const routes = [
     routes: [
       { path: '/', exact: true, component: Home },
       { path: '/about', component: About },
-      { path: '/chat', component: Chat },
       { path: '/shop', exact: true, component: Shop },
       { path: '/shop/:itemid', component: Item },
       { path: '/radio', component: RadioHome },
@@ -61,12 +56,9 @@ const routes = [
       { path: '/boards/:boardtag/:threadid', component: Thread },
       { path: '/discovery', component: DiscoveryHome },
       { path: '/blog', component: BlogHome },
-      { path: '/chat-feathers', component: isAuthenticated(ChatFeathers) },
       { path: '/login', component: Login },
       { path: '/login-success', component: isAuthenticated(LoginSuccess) },
       { path: '/register', component: isNotAuthenticated(Register) },
-      { path: '/survey', component: Survey },
-      { path: '/widgets', component: Widgets },
       { component: NotFound }
     ]
   }
